+++ src/gui/Settings.cc
@@ -79,7 +79,7 @@ void SettingsEntryEnum::setValue(const std::string& value)
 static std::vector<SettingsEntryEnum::Item> axisValues() {
   std::vector<SettingsEntryEnum::Item> output;
   output.push_back({"None", _("None")});
-  for (int i = 0; i < InputEventMapper::getMaxAxis(); ++i) {
+  for (size_t i = 0; i < InputEventMapper::getMaxAxis(); ++i) {
     const auto userData = (boost::format("+%d") % (i + 1)).str();
     const auto text = (boost::format(_("Axis %d")) % i).str();
     output.push_back({userData, text});
@@ -205,7 +205,7 @@ SettingsEntryDouble Settings::axisDeadzone8("input", "axisDeadzone8", 0.0, 0.01,
 SettingsEntryInt Settings::joystickNr("input", "joystickNr", 0, 9, 0);
 
 
-SettingsEntryString& Settings::inputButton(int id)
+SettingsEntryString& Settings::inputButton(size_t id)
 {
   const std::array<SettingsEntryString *, InputDriver::max_buttons> entries {
     &inputButton0,  &inputButton1,  &inputButton2,  &inputButton3,
@@ -219,7 +219,7 @@ SettingsEntryString& Settings::inputButton(int id)
   return *entries[id];
 }
 
-SettingsEntryDouble& Settings::axisTrim(int id)
+SettingsEntryDouble& Settings::axisTrim(size_t id)
 {
   const std::array<SettingsEntryDouble *, InputDriver::max_axis> entries {
     &axisTrim0, &axisTrim1, &axisTrim2,
@@ -230,7 +230,7 @@ SettingsEntryDouble& Settings::axisTrim(int id)
   return *entries[id];
 }
 
-SettingsEntryDouble& Settings::axisDeadzone(int id)
+SettingsEntryDouble& Settings::axisDeadzone(size_t id)
 {
   const std::array<SettingsEntryDouble *, InputDriver::max_axis> entries {
     &axisDeadzone0, &axisDeadzone1, &axisDeadzone2,
